this fails compile due to the lifetime of SES
